packages/data/languages/pyproject.toml
[project]
name = "languages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.3"
dependencies = [
    "pandas==3.0.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
]
